cloud-functions/anomaly-detector/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.15.1

